Add groups and differentiate between forms via form config
