+++ blocks/form/decorators/layout.js
@@ -1,6 +1,6 @@
 function createSection(fields, groupName, form) {
   const section = document.createElement('div');
-  section.className = groupName;
+  section.className = groupName + (form.id ? ` ${form.id}-${groupName}` : '');
   section.append(...form.querySelectorAll(fields));
   return section;
 }
